fm - make sure the datacenters side can handle concurrent requests (thread-safe). Refactor also some parts of the code base.
