+++ replicated-commit/src/main/java/edu/ucsb/rc/Shard.java
@@ -2,6 +2,8 @@ package edu.ucsb.rc;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
 
 import edu.ucsb.rc.locks.LocksManager;
 import edu.ucsb.rc.network.Message;
@@ -15,13 +17,11 @@ public class Shard {
 	private int shardID;
 	private String ipAddress;
 	private Datacenter datacenter = null;
-	private HashMap<String, Transaction> transactionsMap;
 	private LocksManager locksManager;
 	private TwoPhaseCommitManager twoPCmanager;
 	private PaxosAcceptsManager paxosAcceptsManager;
 	
 	public Shard() {
-		this.transactionsMap = new HashMap<String, Transaction>();
 		this.locksManager = new LocksManager();
 	}
 	
@@ -56,7 +56,6 @@ public class Shard {
 	
 	public void handleReadRequestFromClient(Transaction t) {
 		// Handle a read request coming from a client
-		this.addTransaction(t);
 		
 		ArrayList<Operation> readSet = t.getReadSet();
 		boolean allSharedLocksAcquired = true;
@@ -85,7 +84,6 @@ public class Shard {
 					this.locksManager.removeLock(op.getKey(), t.getServerTransactionId());
 				}
 			}
-			this.removeTransaction(t);
 			messageForClient.setMessageType(Message.MessageType.READ_FAILED);
 		}
 		
@@ -95,7 +93,6 @@ public class Shard {
 	
 	public void handlePaxosAcceptRequest(Transaction t) {
 		// Handle a PAXOS Accept request coming from client
-		this.addTransaction(t);
 		
 		NetworkHandler networkHandler = MultiDatacenter.getInstance().getNetworkHandler();
 		
@@ -121,72 +118,31 @@ public class Shard {
 		// Handle a 2PC prepare request
 		Shard shardSender = this.datacenter.getShard(shardIdOfSender);
 		
-		if (!this.checkSharedLocksAreStillAcquiredForTxn(t)) {
+		if (!this.locksManager.checkSharedLocksAreStillAcquiredForTxn(t)) {
 			// Remove all shared locks for Txn
-			this.removeAllSharedLocksForTxn(t);
+			this.locksManager.removeAllSharedLocksForTxn(t);
 			
 			this.sendMessageToOtherShard(shardSender, Message.MessageType.TWO_PHASE_COMMIT__PREPARE_DENIED, t);
 			return;
 		}
-		if (!this.acquireExclusiveLocksForTxn(t)) {
+		if (!this.locksManager.acquireExclusiveLocksForTxn(t)) {
 			// Remove all locks for Txn
-			this.removeAllLocksForTxn(t);
+			this.locksManager.removeAllLocksForTxn(t);
 			
 			this.sendMessageToOtherShard(shardSender, Message.MessageType.TWO_PHASE_COMMIT__PREPARE_DENIED, t);
 			return;
 		}
 		
 		// Remove all shared locks for Txn
-		this.removeAllSharedLocksForTxn(t);
+		this.locksManager.removeAllSharedLocksForTxn(t);
 		
 		this.sendMessageToOtherShard(shardSender, Message.MessageType.TWO_PHASE_COMMIT__PREPARE_ACCEPTED, t);
-	}
-	
-	public boolean checkSharedLocksAreStillAcquiredForTxn(Transaction t) {
-		ArrayList<Operation> readSet = t.getReadSet();
-		
-		for (Operation readOp : readSet) {
-			if (!this.locksManager.isLockedByTransaction(readOp.getKey(), t.getServerTransactionId())) {
-				return false;
-			}
-		}
-		return true;
-	}
-	
-	public boolean acquireExclusiveLocksForTxn(Transaction t) {
-		ArrayList<Operation> writeSet = t.getWriteSet();
-		
-		for (Operation writeOp : writeSet) {
-			if (!this.locksManager.addExclusiveLock(writeOp.getKey(), t.getServerTransactionId())) {
-				return false;
-			}
-		}
-		return true;
-	}
-	
-	public void removeAllSharedLocksForTxn(Transaction t) {
-		ArrayList<Operation> readSet = t.getReadSet();
-		
-		for (Operation readOp : readSet) {
-			this.locksManager.removeLock(readOp.getKey(), t.getServerTransactionId());
-		}
-	}
-	
-	public void removeAllLocksForTxn(Transaction t) {
-		this.removeAllSharedLocksForTxn(t);
-		ArrayList<Operation> writeSet = t.getWriteSet();
-		
-		for (Operation writeOp : writeSet) {
-			this.locksManager.removeLock(writeOp.getKey(), t.getServerTransactionId());
-		}
 	}
 	
 	public void handleTwoPhaseCommitPrepareAccepted(Transaction t, int shardIdOfSender) {
 		/* We signal a new accepted 2PC prepare. If everybody has accepted, we send
 		 * paxos accept accepted to client and all other coordinators in other datacenters.
 		 */
-		
-		NetworkHandler networkHandler = MultiDatacenter.getInstance().getNetworkHandler();
 		
 		if (this.twoPCmanager.signalAcceptedPrepare(t)) {
 			this.twoPCmanager.stopTracking2PCaccepts(t);
@@ -249,27 +205,8 @@ public class Shard {
 			}
 		}
 		
-		this.removeAllLocksForTxn(t);
+		this.locksManager.removeAllLocksForTxn(t);
 		// Transaction is officially committed in this shard!!!
-	}
-	
-	private void addTransaction(Transaction t) {
-		this.transactionsMap.put(t.getServerTransactionId(), t);
-	}
-	
-	private boolean containsTransaction(String serverTransactionId) {
-		return this.transactionsMap.containsKey(serverTransactionId);
-	}
-	
-	private Transaction getTransaction(String serverTransactionId) {
-		return this.transactionsMap.containsKey(serverTransactionId) ? 
-				this.transactionsMap.get(serverTransactionId) : null;
-	}
-	
-	private void removeTransaction(Transaction t) {
-		if (this.containsTransaction(t.getServerTransactionId())) {
-			this.transactionsMap.remove(t.getServerTransactionId());
-		}
 	}
 	
 	public boolean operationKeyBelongsToCurrentChard(Operation op) {
